update routes and models
