+++ App.ts
@@ -10,9 +10,9 @@ import { FavoriteList } from './route/FavoriteList';
 
 //connect to the model 
 import {DataAccess} from './DataAccess';
-import {RestaurantModel} from './model/RestaurantModel';
-import {DishModel} from './model/DishModel';
-import {RestaurantTagListModel} from './model/RestaurantTagListModel';
+import { Restaurant } from './route/Restaurant';
+import { Menu } from './route/Menu';
+import { RestaurantTagList } from './route/RestaurantTagList';
 
 // Creates and configures an ExpressJS web server.
 class App {
@@ -54,6 +54,28 @@ class App {
         review.registerRoutes(router);
         var favoriteList = new FavoriteList();
         favoriteList.registerRoutes(router);
+        // erica
+        this.addRestaurant(router);
+        this.addMenu(router);  
+        this.addrTags(router);
+    }
+
+    /******** Restaurant ********/
+    private addRestaurant(router: express.Router): void{
+      var rest = new Restaurant();
+      rest.registerRestaurantRoutes(router);
+    }
+    
+    /******** Restaurant Dish********/
+    private addMenu(router: express.Router): void{
+      var menu = new Menu();
+      menu.registerDishRoutes(router);
+    }
+
+    /******** Restaurant Tags********/
+    private addrTags(router: express.Router): void{
+      var rtaglist = new RestaurantTagList();
+      rtaglist.registerrTagListRoutes(router);
     }
 
 }
