+++ pages/StyleSheet.css
@@ -5,6 +5,7 @@ body{
   font-family: 'Open Sans', sans-serif;
 }
 
+/*search*/
 .search {
   width: 100%;
   position: relative;
